Drop support for go pre 1.10
